(PC-35362) refactor(query): trusted device
